chore: use http link for github repo
The page for this module on npm [doesn't have a link to the repo](https://www.npmjs.com/package/libp2p-noise),
I guess because it's specified as an ssh-type URL in the package.json
file so npm can't turn it into a hyperlink.

The change here is to use a `git+https` URL which should let npm
display a link on the module page.