fix!: update multiformats and related dependencies
`multiformats@11.x.x` shipped with a [breaking change](https://github.com/multiformats/js-multiformats/pull/230)
so update all deps using multiformats to the latest version.

The code changed is not used by this module so there's not a lot of
changes here, but this should still go out as a major because the
interface versions have changed.