feat: add service capabilities definition
The latest libp2p release allows services to declare their
capabilities (what they provide to the node) and their
dependencies (the capabilities they require other services to
provide for the config to be "valid").

Adds the `serviceCapabilities` symbol property that says Noise
provides the `@libp2p/connection-encryption` capability and also
that it provides noise encryption as some transports (WebTransport,
WebRTC) now require this module to be configured alongside them.